Make sure networks_skip_config defaults to []
The ansible network configs does a check if network
is in networks_skip_config. When networks_skip_config is
not defined in the role data it get's defined as 'null'
in the Ansible inventory.

Default it to an empty list when creating the GroupVars
so that it's always of type: List in the inventory.

Change-Id: I0c09118b947aeee9b011e7d0ec23ab99af3680af
Closes-Bug: #1904808
(cherry picked from commit c8c9855678540ffb77734d46d7dd76f2b58ebf94)