Remove corosync.conf if it's a dir from remote.
During the upgrade from Queens there can be some wrong config
remainings. One of them is the existence of /etc/corosync/
corosync.conf as a directory. This will cause pacemaker to fail
when the node is upgrade to Train, failing with:
Error:Unable to read /etc/corosync/corosync.conf: Is a directory.

This patch checks during the upgrade if the file exists and it is
a directory. In that case, it will get removed.

Change-Id: I26f9fea0a0495e5e976a56a50c70a15e47d14400
Closes-Bug: #1905033
Resolves: rhbz#1899935