Remove vfio-pci.conf module load file
With an existing BZ #1898664 on dracut does not create ramfs with
vfio_iommu_type1 module, because which loading vfio-pci during the initramfs
fails to load this module. Because of this dpdk ports are added in ERROR
state. It requires a restart ovs to bring to normal state after ffu is
complete. As a workaround, the module-load file vfio-pci.conf is removed
before upgrade, which will ensure that vfio-pci is not loaded during initramfs
and it will be loaded when driverctl configures the vfio-pci driver to the
interface.
Closes-Bug: #1905533

Change-Id: I752a764a53e90fcb17e414d4900bb186fa689f45
(cherry picked from commit a487e16c55fc4346e250f41bea4ee8fb43de55c5)