Set correct default NovaLibvirtCPUMode
https://review.opendev.org/q/I8df21d5d171976cbb8670dc5aef744b5fae657b2
introduced THT parameters to set libvirt/cpu_mode. The patch sets the
NovaLibvirtCPUMode wrong to 'none' string which results in puppet-nova
not to handle the default cases correct and sets libvirt/cpu_mode to
none which results in 'qemu64' CPU model, which is highly buggy and
undesirable for production usage.  This changes the default to the
recommended CPU mode 'host-model', for various benefits documented
elsewhere.

Closes-Bug: #1905544

Change-Id: Iea8cccd77caac4b84764d84a213918ed57bd4e3e