Enable debug logging of libvirt services when Debug is true
This patch ensures that debug logging of libvirt services is enabled
when the global Debug parameter is true, since the parameter is
supposed to enable debug logging for all services.

Also it introduces a new LibvirtDebug parameter, to allow operators to
enable debug of libvirt services more easily, than setting log level
number between 1 and 4.

Change-Id: I54e0ee6fe59d04686f15cb5638262b34572596e1