Add qemu metadata to compute node when tls for live migration
With I7f583d18e558b95922a66eb539cc91de74409c96 certificates are
moved to use bind mounts and in case of UseTLSTransportForNbd
to create the required certificates even if UseTLSTransportForNbd
is set to False. In case of UseTLSTransportForNbd is False the
compute node still need the qemu metadata to have permissions
to request the certificates for the nbd tls use case.

Change-Id: Ibba61afbeb3957a955aa6d75e8258279a60fd141