Do not relabel Swift files on every container start
Without this patch, files in /srv/node are relabeled on every start of
the account_auditor and/or account_reaper containers. If there are many
files (eg. when using Gnocchi) this will take a long time, sometimes
dozens of minutes per container start, and might result in breaking
upgrades/updates.

Relabeling already happens in step 3, this should be sufficient and
prevent additional delays when (re-) starting containers.

Closes-Bug: 1907070
Change-Id: I172ae8f35df34887aaf61b3e03d5aaab1d462a60