[Ussuri and older] Set python_cmd where we need it
If you use --start-at-tasks, python_cmd may not be defined when we run
the generate configuration or host container tasks.  We should ensure
this variable is set prior to using it.

This variable was no longer used with the transition of the
container-puppet.py items into ansible itself so it only applies for
Ussuri and older.

Change-Id: I956307d5c69cd4d259fb1ab2731fe0001a196c8e
Closes-Bug: #1907858
(cherry picked from commit 1a085631c09d3617c60ca2f5b18eca9f1385aae5)