Rely on the HOSTNAME var to resolve the mon container name
We previously used the {{ ansible_hostname }} variable as a replacement
for ${HOSTNAME} when the ceph-mon container name needed to be resolved
during the FFU process.
However, sounds like {{ ansible_hostname }} doesn't return the value of
the delegated host [1].
The purpose of this change is to compute the hostname relying on the
exported env var and it's supposed to fix the tls-everywhere use case
(when a more complex domain is provided).

[1] https://github.com/ansible/ansible/issues/14958

Closes-Bug: #1903537
Change-Id: I77e3e5e4336a5d3f76b10d09cf7034023fbec075