tool: convert heat network-config to ansible j2
Add a script that does a best effort conversion of
a heat network-config tempalte to an ansible j2
network-config template.

The script uses a networks data file as input to
genereta a map of Heat Parameters to ansible vars.
For parameters not in the generated map the stack
environment is used, parameter values from the
stack environment is hard-coded in the j2 template.

A j2 comment is added whenever a value was hard coded,
in a header comment and also an inline comment if
possible.

NOTE: The j2 reference files in the unit tests was
      created by converting heat templates.

Change-Id: I8165a077b87307ca3c2ebee54703a939517dc9bf