Remove duplicated tasks to create persisitent directries for haproxy
We already have the task to create persistent directories for haproxy
in the HAProxyBase resource, so we don't need to define the same task
in haproxy-pacemaker-puppet.yaml.

Note that haproxy-pacemaker-puppet.yaml used to create /var/log/haproxy
but this is removed by this change because that directory is never
used.

Change-Id: I8a846317512ad068caa297821105023a71dc2183