[Queens] Handle ceph service restart during update.
Since this change[1] in ceph-ansible the ceph services are stopped if
docker is stopped. This happens during the update steps when docker
needs to be updated[2].

This in turn breaks the update in several ways.

To avoid any problem, when docker need a restart, we:
 1. make sure we restart the ceph services if docker was stopped;
 2. set the necessary flags so that no rebalance action happens when
    osd services are stopped.

To achieve this we make sure to set the docker restart related facts
in steps 0 so that they are available for step1,2,4.

For the flag we also set "run_once" so that setting the flags happens
only once if the ceph-osd are updated in batch[3].

I resuffle the when conditional on top of the blocks to work around
lp#1910418.

[1] https://github.com/ceph/ceph-ansible/pull/5444
[2] https://opendev.org/openstack/tripleo-heat-templates/src/branch/stable/queens/puppet/services/docker.yaml#L178-L184
[3] Shouldn't be the default as osd should be updated serially, but it's cleaner this way.

Queens-Only because it happens only when using docker.

Change-Id: If7ebd5b0fed0c84b720729e1e75f659bae2355fe
Closes-Bug: #1910124