Revert "Reset sriov_numvfs to 0 before leapp upgrade"
This reverts commit b3ec034649e0d18de78fcfe5f903cfe9a0a18a6a.

Reason for revert: The patch was introduced as a workaround
to fix the leapp issue reproduced in v0.10.0 which throws
exception when the interface is not found. But with the
recent version of the leapp package, the issue is not
reproducible. The leapp upgrade completes succesfully
even with VFs present. Though the actual leapp commit which
fixed the issue not traced, the issue is no more present.
This workaround can be removed from TripleO, so that
nic paritioning FFU can be supported.

This revert is required for nic paritioning FFU, because
resetting the VFs will hang, when one of the VF is bound
to the vfio-pci (DPDK) driver.

Change-Id: I5ab1d2925989fc3da62e6045e56a7bd017c8ec4c
(cherry picked from commit c4d75bc147d0e5fe56b9445d23f0a1ea7a34b141)
(cherry picked from commit 81a18403304f6f7dbb1a58be68875cc809ad28c4)