Force json output format for hiera in derive pci whitelist
For some reason, after queens to train FFU, when drive pci
whitelist is executed, the hiera output is shown with the
default output format 'ruby' (with => instead of :). This
results in json parsing error. Force the output to be json
which will work for regular and the ffu cases.

Change-Id: I065ec0e7ae14530f6bf4f6fe356e885ba2b30d7f
(cherry picked from commit 8a79c1b635efb545bdd23c2ecc6fbe11c34bd924)