nova: Use LIBGUESTFS_BACKEND=direct
After change [1] nova-compute launch libguestfs using the default
``qemu:///system``, but when ``inject_password` is set to true and
user tries to create vm, the vm creation is successful and we could
see libguestfs error in nova-compute logs.

This change forces libvirt to use ``direct`` when launching instances
on host.

[1] Ib55936ea562dfa965be0764647e2b8e3fa309fd6

Change-Id: I195358742c19d6ea0a3d32979896c0268e3b55a6
Closes-bug: #1912141