[Train only] Retry distro-sync until success
The distro-sync process is quite lengthy and very network heavy
and it is crucial that it completes. If there are network comms
issues or package serving issues from the repository serving
them then it will fail.

To increase the chances of success, we add 5 retries with a 10
second delay between them to allow some time for the repository
to recover.

Change-Id: Ia6421bd534f8355d79aab48cb4146d067bdd61b0