Deprecate environments/dcn-hci.yaml for dcn-storage.yaml
Rename the file in the environments directory so that it
reflects its expanded scope. This file is used when
deploying storage with DCN sites regardless of if those
sites use HCI. We are now supporting non-HCI DCN sites
with storage so the old name is confusing.

Old name : dcn-hci.yaml
New name : dcn-storage.yaml

dcn-hci.yaml is depreacated but will remain in the environments
directory for backwards compatibility. dcn-hci.yaml will be
removed during the X cycle.

Change-Id: Ice5e1cfbc158eb6705988706c8625bedb80d7de2