Use include_role for conditional inclusion
When import_role is used with a condition, the condition is
applied to all tasks in the role. This is inefficient. If we
use include_role instead, then the role inclusion task is
skipped and none of the tasks in the role are even evaluated.

Related-bug: rhbz#1922132

Change-Id: I1b4e26fd4a46599985a989441f493a3ed39237bb
(cherry picked from commit 46df551a0f00cf143ddd5cd9c651b08a13a79fb0)