Split network validation to it's own play
When we moved the network deployment to free, we can hit a race
condition where some nodes are being configured prior to the controllers
being ready. This can lead to the basic network validation failing.  We
can deal with this by moving the validation to it's own play so that we
ensure all network configurations have occurred prior to to checking
that nodes can ping the controllers.

Change-Id: I48665379a87e701633be1bcc90bd8be1cc75c513
Closes-Bug: #1913725
(cherry picked from commit 64e735898b8fae7643fc092fed840d867a252046)