Hard code jobs to Nautilus before setting new default to Octopus
The jobs being changed in this patch would break when Ocotpus
becomes the default Ceph container for the future Wallaby [1][2].
This patch configures these jobs to use Nautilus as an exception,
not the rule, so that these jobs may then be migrated to Octopus
one by one (i.e. migrated to deploying Ceph via cephadm). This
patches change for 001-standalone will be undone first via [3].

[1] https://review.opendev.org/c/openstack/tripleo-common/+/771424
[2] https://review.opendev.org/c/openstack/tripleo-quickstart-extras/+/772371
[3] https://review.opendev.org/c/openstack/tripleo-heat-templates/+/767294

Change-Id: If9d51a2412c6b20ca3d86d2905831b754cc0622f