[train-only] Ensure we stop ovn-controller with cleanup
Turns out that gracefull shutdown of container running ovn_controller
does not disable routing to it. This means that after Leapp when the
node is clean we still try to send traffic to it and this causes
outage. By calling ovs-appctl -t ovn-controller exit we ensure that
ovn cleans up BFD interfaces from the node. For this to work we need
to disable docker autorestart of the container.

Change-Id: I9ec73e3d94884c339ea5e21817c24cf5e758f242