set ganesha idmap.conf file path and overrides
Traditionally,  Ganesha consumes /etc/idmapd.conf, which controls
mapping of user/owner identities under NFSv4+.  With containerized service
deployment, this file has beenan immutable part of the container image
and could not be modified.

Recently ceph-ansible was modified to allow the user to set the path
for idmapd.conf and to override ini settings in this file by supplying
json configuration for the overrides in the same way that it can be
supplied for ceph.conf overrides [1].  Also, the default path setting
for the idmapd.conf file is /etc/ganesha/idmap.conf, which already is
reflected into containerized ganesha under /etc/ganesha, solving the
immutability issue.  If no overrides are supplied, the file put in this
location will match that originally frozen into the container, so there
is no issue here with backwards compatibility.

This change exposes parameters for idmap.conf file path and override
contents in TripleO and passes them into ceph-ansible.  It also sets up
these parameters for cephadm.  Cephadm support for OpenStack Ganesha
is WIP so there will likely need to be a followup patch to make these
effective when cephadm replaces ceph ansible.

[1] https://github.com/ceph/ceph-ansible/pull/6422

Change-Id: I422e49f0027efbbb25256d8b4718eb0aa78d90a4