Check Ceph cluster healthy state before starting FS to BS playbook
The Ceph OSDs migration from filestore to bluestore should only be
attempted if the cluster is in HEALTH_OK state because during the
process OSDs will be destroyed and recreated.

This change adds a check to poll for the cluster status before the
playbook is triggered.

Change-Id: Ib9c90c4c2e838bed461ac63139b161ab689d28cf
Depends-On: https://review.opendev.org/777561
Closes-Bug: 1915700
(cherry picked from commit 5250e6d595c7ad97059930809fbeda9b17c1b364)