pcs commands on host: redis bundle
This implements the creation of the redis bundle on the host.
The testing protocol used is documented in the depends-on.

The full rationale is contained in the LP bug.

The reason for adding a post_update task is that during a minor update
the deployment tasks are not run during the node update procedure but
only during the final converge. So we ran the role again there to make
sure that any config change will trigger a restart during the minor
update, so the disruption is only local to the single node being
updated. If we did not do this a final converge could potentially
trigger a global restart of HA bundles which would be quite disruptive.

Related-Bug: #1863442

NB: Conflicts mainly due to context diffs

Depends-On: Iaa7e89f0d25221c2a6ef0b81eb88a6f496f01696

Change-Id: I5ce8367363d535b71b01395b0bef4cf17c8935b5
(cherry picked from commit 83d1cb585268177ef6f685e689d5aa34b24c8056)
(cherry picked from commit 0eb998930011ee3fc13ecc188dbeb578f185b1ec)