pcs commands on host: mysql
This implements the creation of the haproxy bundle on the host.
The testing protocol used is documented in the depends-on.

The reason for adding a post_update task is that during a minor update
the deployment tasks are not run during the node update procedure but
only during the final converge. So we ran the role again there to make
sure that any config change will trigger a restart during the minor
update, so the disruption is only local to the single node being
updated. If we did not do this a final converge could potentially
trigger a global restart of HA bundles which would be quite disruptive.

NB: in this case we keep the container init_bundle (renamed to
wait_bundle) around just use it to wait for galera to be up.

Depends-On: Iaa7e89f0d25221c2a6ef0b81eb88a6f496f01696

NB: We also merged Ic6c65e6849368185177aeaa31d50f52761225f62
    in this review so that the ussuri upgrade job can pass
    (Make sure we noop the Mysql_ providers)

NB: Some conflicts due to context differences

Change-Id: Ie14819b66cecdb5a9cc6299b68a0cc70a7aa3370
Related-Bug: #1863442
(cherry picked from commit da3d5e80562f7aa3d0039d44a2fd0a2f4c093cf4)
(cherry picked from commit be1fee4af6a5ed7ef4eff4ede26f5d3666132398)