Swift: Enable gatekeeper and listing_formats explicitly.
The gatekeeper middleware and the listing_formats middleware are
required in swift-proxy-server pipeline and automatically enabled even
if the middlewares are not explicitly enabled by the pipeline parameter.

This change ensures that these two middlewares appear in the list
passed to the pipeline parameter so that operators can more easily
understand which middleware is actually enaled.

Depends-on: https://review.opendev.org/776826
Depends-on: https://review.opendev.org/776830
Change-Id: Iedde84563aff3def0ec20386399eea7f244cf075