Add delegate_fact_hosts: false on ci scenarios
This change adds delegate_facts_hosts: false to the
existing ceph-ansible scenarios.
This was introduced due to the --limit option to avoid
gathering facts, but since we're running on standalone
and the same node is present in all groups, having that
variable set to True (which is the ceph-ansible default)
makes no sense.

Change-Id: I44433731f73882f62591e8067743beec4d423ef7
(cherry picked from commit c0e8697550bba624bd935d5ad01135a1d10094e9)