Adds net_cidr_map variable to allow cidr lists
This patch exposes the net_cidr_map variable so that tasks can
access the list of CIDRs that are valid for a network as opposed
to attempting to build the CIDRs from the network definitions.

In spine-leaf or edge use cases the networks may have multiple
subnets assigned to a given network.

The new Unbound service will use these maps to build lists of
CIDRs allowed to make queries.

Change-Id: I6004519e8b2317d19356c4a2b8bea416b4d94c22