Add CephDynamicSpec, CephSpecPath, CephOsdSpec, CephSpecFqdn
If the CephDynamicSpec boolean is true, then the
ceph_spec_bootstrap Ansible module will create a
Ceph spec file based on the tripleo-ansible-inventory.

If the CephDynamicSpec boolean is false, then the
deployer may place their own Ceph spec file on the
undercloud at CephSpecPath.

If CephOsdSpec json is set, then it will appear in the
genereated Ceph spec for the 'osd' service_type.

If CephSpecFqdn boolean is true, then the hostname
and hosts of the generated Ceph spec will have their
fully qualified domain name instead of their short
hostname.

Depends-On: I74cb373f1dc94f8a5ef12c0545cfbc0c49abd69b
Change-Id: I3ce3db44706c4b378aa5abc7650ad8478576bf3a