Introduce CephHciOsdCount and CephHciOsdType
CephHciOsdCount is the number of expected Ceph OSDs per HCI node.
CephHciOsdType is the type of data_device (not db_device) used for
each OSD and must be one of hdd, ssd, or nvme. These are used by
the Ansible module tripleo_derive_hci_parameters. Since CephOsdSpec,
as used by cephadm, might only specify a description of devices to
be used as OSDs (e.g. all devices), and not a list of devices like
CephAnsibleDisksConfig, setting the count directly is necessary in
order to know how much CPU/RAM to reserve. Similarly, because a
device path is not hard coded, we cannot look up that device in
Ironic to determine its type.

Closes-Bug: #1920954
Depends-On: Ia6bbdf023e2a0961cd91d3e9f40a8a5a26253ba3
Change-Id: Iccf97ca676ee6096e47474c571bd4f53381ce1c9