HA: fix race when moving VIP during minor update
During a minor update of a controller node, we list the
VIP hosted on the node and we force them to move away
with a pcs command.

The way we call the pcs command is racy, so change the
action to make sure the VIP move is performed only if
the VIP is still hosted on the node when we run the action.

Change-Id: Id379f4fe1668d01fdd5f91b46e2f75d7cdb577ae
Closes-Bug: #1921351
(cherry picked from commit 3da1e7661036a73c9f222bcdcc489912502d23cb)
(cherry picked from commit de98fdb208dac1cf01d474c4363f6e2177e22c7e)