Remove CrushRules coverage from scenario001
We see many issues in CI due to the missing hdd device class.
This is something that can be related to the fact that only
1 OSD in this job is built on top of a loop device.
For this reason, and in order to keep the CI healthy, it's
safe removing the device classes coverage.
In addition, since there is only 1 osd, add and configure
tiering in this context makes no sense.

Change-Id: I4638b9badd17a7a75642895623a63753534fd341