Fix tls for undercloud with ipa
We recently changed cert generation to use linux-system roles
to generate certs instead of puppet-certmonger.  However, this broke
the ability to generate the haproxy cert on the undercloud using an
IPA server, because we relied in the ability to specify the CertmongerCA
and the hieradata to provide the correct ca, principals and dns entries.

This patch restores this ability through THT template parameters.

Change-Id: Ie2e181fcd9198ae5613fde7135230d4b4cf7343d