Add OS::TripleO::UndercloudUpgradeEphemeralHeat
A new service, OS::TripleO::Services::UndercloudUpgradeEphemeralHeat is
added to the Undercloud role. The service is mapped to OS::Heat::None by
default, but when environments/lifecycle/undercloud-upgrade-prepare.yaml
is included, the service will be enabled and will migrate any already
deployed stacks in the undercloud's Heat instance to be able to be used
with the ephemeral Heat deployment option from tripleoclient.

Signed-off-by: James Slagle <jslagle@redhat.com>
Change-Id: If11e2fc07a1ff773f6eaf209d8b48493f0b60e85