[ffwd][train-only] Run keystone endpoint configuration on FFWD
By default during FFWD we run all tasks with --limit and we never
specify Undercloud. This means we never run the
external_post_deployment_tasks which sets up the keystone
endpoints. This is intended as we don't want to trigger most of
external deployment tasks at that point. Solution to missing
endpoints is to run the keystone setup during post_upgrade_tasks.

Change-Id: Ic1e63ca773fb20f10f16fc3835f8f263cb3b6929
Resolves: rhbz#1944108